Changed GetMilliseconds() to return a 64-bit value
This function previously returned an int32_t. Presuming that it measures
the length of time since the application was started, the 32-bit type would
cause the returned value to wrap from 2^31-1 to -2^31 after a little less
than twenty-five days.

+++ graphicswin.cpp
@@ -263,7 +263,7 @@ void GraphicsWindow::AnimateOnto(Quaternion quatf, Vector offsetf) {
     // long translations (as measured in pixels) if the user zooms out, moves,
     // and then zooms in again.
     if(dt > 2000) dt = 2000;
-    int32_t tn, t0 = GetMilliseconds();
+    int64_t tn, t0 = GetMilliseconds();
     double s = 0;
     Quaternion dq = quatf.Times(quat0.Inverse());
     do {
